hiring request field updated

+++ HumanitarianAssistance.Application/Project/Commands/Create/AddProjectHiringRequestCommand.cs
@@ -10,7 +10,7 @@ namespace HumanitarianAssistance.Application.Project.Commands.Create {
         public int? Office { get; set; }
         public int? GradeId { get; set; }
         public long? ProjectId { get; set; }
-        public long? JobCategory { get; set; }
+        public int? JobCategory { get; set; }
         public string MinEducationLevel { get; set; }
         public string Organization { get; set; }
         public int? ProvinceId { get; set; }
