+++ HumanitarianAssistance.Application/Project/Models/ProjectHiringRequestDetailModel.cs
@@ -5,7 +5,7 @@ namespace HumanitarianAssistance.Application.Project.Models {
         public long? HiringRequestId { get; set; }
          public string HiringRequestCode { get; set; }
         public long? ProjectId { get; set; }
-        public long? JobCategory { get; set; }
+        public int? JobCategory { get; set; }
         public string MinEducationLevel { get; set; }
         public int? TotalVacancy { get; set; }
         public int? Position { get; set; }
